requirements.txt
beautifulsoup4==4.4.1
capstone==3.0.4
click==6.6
construct==2.5.2
Demangler==0.1
enum34==1.1.6
Flask==0.11.1
Flask-Assets==0.11
gunicorn==19.6.0
hurry.filesize==0.9
itsdangerous==0.24
Jinja2==2.8
lxml==3.6.0
MarkupSafe==0.23
pathlib==1.0.1
-e git+https://github.com/eliben/pyelftools.git@5d3b1545df30d4bc3a6cfa8741b1f08cb6bf204b#egg=pyelftools
pyparsing==2.1.5
pyScss==1.3.5
six==1.10.0
webassets==0.11.1
Werkzeug==0.11.10
